% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fev_stack.R
\name{as_fev_stack}
\alias{as_fev_stack}
\title{Coerce to a \code{fev_stack}}
\usage{
as_fev_stack(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Passed to \code{\link[=fev_stack]{fev_stack()}}.}
}
\value{
A \code{fev_stack}.
}
\description{
Coerce to a \code{fev_stack}
}
